// src/app/models/forecast-to-plot.ts
// Generated by https://quicktype.io

import { TruthToPlotSource, TruthToPlotValue } from './truth-to-plot';

export enum ForecastToPlotType {
  Observed = 'observed',
  Point = 'point',
  Quantile = 'quantile'
}

export interface ForecastToPlotTarget {
  time_ahead: number;
  time_unit: 'day' | 'wk';
  value_type: TruthToPlotValue;
  end_date: moment.Moment;
}

export interface ForecastToPlot {
  forecast_date: moment.Moment;
  target: ForecastToPlotTarget;
  location: string;
  type: ForecastToPlotType;
  quantile: string;
  value: number;
  timezero: moment.Moment;
  model: string;
  truth_data_source: TruthToPlotSource;
  shift_ECDC: number;
  shift_JHU: number;
  first_commit_date: moment.Moment;
}